requirements.txt
# ============================================
# FILE: requirements.txt
# ============================================
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
scikit-learn==1.3.0
Pillow==10.0.0
PyPDF2==3.0.1
python-docx==0.8.11
fpdf==1.7.2
openpyxl==3.1.2